try int64

+++ surf_seis/vibbox.py
@@ -214,7 +214,7 @@ def vibbox_read(fname, param):
     with open(fname, "rb") as f:
         f.seek(HEADER_OFFSET, os.SEEK_SET)
         # read header
-        H = np.fromfile(f, dtype=np.int64, count=HEADER_SIZE)
+        H = np.fromfile(f, dtype=np.int32, count=HEADER_SIZE)
         BUFFER_SIZE=H[0]
         FREQUENCY=H[1]
         NUM_OF_BUFFERS=H[2]
